+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `alarm`. ([#830](https://github.com/nix-rust/nix/pull/830))
 - Added interface flags `IFF_NO_PI, IFF_TUN, IFF_TAP` on linux-like systems.
   ([#853](https://github.com/nix-rust/nix/pull/853))
+- Added `statvfs` module to all MacOS and Linux architectures.
+  ([#832](https://github.com/nix-rust/nix/pull/832))
 
 ### Changed
 - Display and Debug for SysControlAddr now includes all fields.